// src/types/api.ts
export enum ChannelTypes {
    "GUILD_TEXT" = 0,
    "DM" = 1,
    "GUILD_VOICE" = 2,
    "GROUP_DM" = 3,
    "GUILD_CATEGORY" = 4,
    "GUILD_ANNOUNCEMENT" = 5,
    "ANNOUNCEMENT_THREAD" = 10,
    "PUBLIC_THREAD" = 11,
    "PRIVATE_THREAD" = 12,
    "GUILD_STAGE_VOICE" = 13,
    "GUILD_DIRECTORY" = 14,
    "GUILD_FORUM" = 15,
    "GUILD_MEDIA" = 19,

}
// ...
export enum MessageComponentTypes {
    "ACTION_ROW" = 1,
    "BUTTON" = 2,
    "STRING_SELECT" = 3,
    "USER_SELECT" = 5,
    "ROLE_SELECT" = 6,
    "MENTIONABLE_SELECT" = 7,
    "CHANNEL_SELECT" = 8,
    "SECTION" = 9,
    "TEXT_DISPLAY" = 10,
    "THUMBNAIL" = 11,
    "MEDIA_GALLERY" = 12,
    "FILE" = 13,
    "SEPARATOR" = 14,
    "CONTAINER" = 17,
}
type MessageComponent = {
    type: MessageComponentTypes.ACTION_ROW;
    components: MessageComponentActionRow[];
}
type MessageComponentActionRow = MessageComponentButton | MessageComponentSelect;

export type MessageComponentEmoji = {
	id?: string;
	name?: string;
	animated?: boolean;
}